+++ portals-ui/sites/ngs-portal/src/shared/ui/index.ts
@@ -1,1 +1,2 @@
 export { SelectFilter } from './select-filter';
+export { ItemLayout } from './item-layout';
