+++ portals-ui/sites/ngs-portal/src/shared/ui/item-layout/index.ts
@@ -1,0 +1,1 @@
+export { ItemLayout } from './item-layout';
