+++ portals-ui/sites/ngs-portal/src/shared/ui/item-layout/item-layout.tsx
