+++ portals-ui/sites/ngs-portal/src/shared/ui/item-layout/layout-card.tsx
